Embed tests in script
