+++ installer_builder2.py
@@ -197,3 +197,77 @@ def format_data_file(item: Union[str, pathlib.Path]) -> str:
 def _format_nuitka_datafiles(items):
     """Convert a list of data file specifications into Nuitka format."""
     return [format_data_file(item) for item in items]
+
+from unittest import TestCase
+
+class TestDataFileFormatter(TestCase):
+    def test_format_data_file(self):
+        # Test single file
+        self.assertEqual(
+            format_data_file("file.txt"),
+            f"{os.path.abspath('file.txt')}=file.txt"
+        )
+        
+        # Test directory
+        self.assertEqual(
+            format_data_file("testdir/"),
+            f"{os.path.abspath('testdir')}=testdir/=**/*"
+        )
+        
+        # Test pattern
+        self.assertEqual(
+            format_data_file("dir/*.txt"),
+            f"{os.path.abspath('dir')}/*.txt=dir/"
+        )
+        
+        # Test pre-formatted string
+        self.assertEqual(
+            format_data_file("source=target"),
+            "source=target"
+        )
+
+    def test_deployment_flag(self):
+        # Create builder with deployment enabled
+        builder = InstallerBuilder("TestApp", enable_deployment=True)
+        self.assertTrue(builder.enable_deployment)
+        
+        # Create builder with deployment disabled (default)
+        builder = InstallerBuilder("TestApp")
+        self.assertFalse(builder.enable_deployment)
+
+    def test_installer_builder_config(self):
+        
+        builder = InstallerBuilder(
+            "TestApp",
+            version="1.2.3",
+            company_name="Test Company",
+            console=True
+        )
+        
+        self.assertEqual(builder.app_name, "TestApp")
+        self.assertEqual(builder.version, "1.2.3")
+        self.assertEqual(builder.company_name, "Test Company")
+        self.assertTrue(builder.console)
+
+    def test_data_file_handling(self):
+        
+        builder = InstallerBuilder("TestApp")
+        builder.data_files = ["test.txt", "data/*.dat"]
+        builder.data_directories = ["assets/"]
+        
+        # Test that data files are properly formatted
+        formatted_files = _format_nuitka_datafiles(builder.data_files)
+        self.assertTrue(any("test.txt" in f for f in formatted_files))
+        self.assertTrue(any("*.dat" in f for f in formatted_files))
+
+    def test_platform_specific(self):
+       
+        builder = InstallerBuilder("TestApp")
+        
+        if platform.system() == "Windows":
+            # Test Windows-specific paths
+            self.assertTrue(str(builder.dist_path).endswith('dist'))
+        elif platform.system() == "Darwin":
+            # Test macOS-specific paths
+            self.assertTrue(str(builder.dist_path).endswith('dist'))
+
