--- test_main.py
@@ -1,5 +1,0 @@
-# This file will cause Nuitka to error when trying to compile
-import non_existent_module
-
-def main():
-    non_existent_module.some_function()
